obu: parse uv_mult etc. as signed integers
The spec subtracts the signed offset from all of these when using them,
like it does for e.g. ar_coeffs_y_plus_128, although for some reason
the naming scheme is inconsistent here. Either way, it makes more sense
to treat them as signed integers than unsigned integers.

To avoid confusion since the name of the field is the same as the one in
the spec, we mark the type as int8_t (resp. int16_t for the 9-bit field)
to make it clear to the user that these are already signed integers.

+++ src/levels.h
@@ -403,9 +403,9 @@ typedef struct Av1FilmGrainData {
     int8_t ar_coeffs_uv[2][25];
     int ar_coeff_shift;
     int grain_scale_shift;
-    int uv_mult[2];
-    int uv_luma_mult[2];
-    int uv_offset[2];
+    int8_t uv_mult[2];
+    int8_t uv_luma_mult[2];
+    int16_t uv_offset[2];
     int overlap_flag;
     int clip_to_restricted_range;
 } Av1FilmGrainData;
